LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-Document-write-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

Hi, I want to be transformed!

PASS document.write with html assigned via policy (successful URL transformation).
PASS document.writeln with html assigned via policy (successful URL transformation).
FAIL `document.write(string)` throws assert_throws_js: function "_ => {
      document.write('A string');
    }" did not throw
FAIL `document.writeln(string)` throws assert_throws_js: function "_ => {
      document.writeln('A string');
    }" did not throw
FAIL `document.write(null)` throws assert_throws_js: function "_ => {
      document.write(null);
    }" did not throw
FAIL `document.writeln(null)` throws assert_throws_js: function "_ => {
      document.writeln(null);
    }" did not throw
FAIL `document.write(string)` observes default policy assert_equals: expected "Quack, I want to be a duck!" but got "Hi, I want to be transformed!"
FAIL `document.writeln(string)` observes default policy assert_equals: expected "Quack, I want to be a duck!" but got "Hi, I want to be transformed!"
